agents/a2a_agents/samples/roll_dice_agent/agent.py
import random

from google.adk.agents.llm_agent import Agent
# from google.adk.tools.example_tool import ExampleTool
from google.genai import types
from google.adk.agents.remote_a2a_agent import AGENT_CARD_WELL_KNOWN_PATH
from google.adk.agents.remote_a2a_agent import RemoteA2aAgent
from google.adk.a2a.utils.agent_to_a2a import to_a2a
# --- Roll Die Sub-Agent ---
async def roll_die(sides: int) -> int:
  """Roll a die and return the rolled result."""
  return random.randint(1, sides)


root_agent = Agent(
    model='gemini-2.0-flash',
    name="roll_agent",
    description="Handles rolling dice of different sizes.",
    instruction="""
      You are responsible for rolling dice based on the user's request.
      When asked to roll a die, you must call the roll_die tool with the number of sides as an integer.
    """,
    tools=[roll_die],
    generate_content_config=types.GenerateContentConfig(
        safety_settings=[
            types.SafetySetting(  # avoid false alarm about rolling dice.
                category=types.HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT,
                threshold=types.HarmBlockThreshold.OFF,
            ),
        ]
    ),
)

# Load A2A agent card from a file
a2a_app = to_a2a(root_agent, port=8002, agent_card="./agent.json")