agents/a2a_agents/samples/roll_dice_agent/pyproject.toml
[project]
name = "rolldice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ 
    "google-adk[a2a]>=1.15.1",
    "python-dotenv>=1.1.1"
    ]
